# doc/EZshopBackEndApplication_swagger.yaml
openapi: 3.0.3
info:
  title: EZShop Back End Application API
  version: 1.0.0
  description: |
    OpenAPI definition for the EZShop backend server application.

    **EZShop Description**

    EZShop is a retail management system designed to handle the end-to-end operations of a small to medium-sized shop, integrating user management, inventory control, sales, orders, and accounting functions into a single platform.
    EZShop defines a set of business operations that orchestrates user interactions, product lifecycle, and financial transactions.

    In a Python-based reimplementation using FastAPI, EZShop would act as a RESTful backend service, exposing endpoints to manage authentication, product catalog, sales transactions, customer information, and payments.
    Authentication would rely on JWT-based bearer tokens, with access permissions differentiated by user roles (Administrator, ShopManager, Cashier).
    
    **Hierarchical Structure and Components**

    The system architecture is logically divided into four layers:

    1. Users and Authentication Layer
       - Manages user accounts and roles (Administrator, Cashier, ShopManager).
       - Provides endpoints for login, logout, and user CRUD operations.
       - Handles session security and authorization for all operations.

    2. Inventory and Product Management Layer
       - Handles product types, quantities, and barcodes (with optional RFID tracking).
...

...

...
      
...
    
...

...

...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - name
      properties:
        code: 
          type: integer
          description: HTTP status code of the error
        message:
          type: string
          description: message of the error
        name:
          type: string
          description: Descriptive error message
      description: Standard structure for error responses
...
    BooleanResponse:
      type: object
      properties:
        success: 
          type: boolean
      required: [success]
    OrderStatus:
      type: string
      enum: [ISSUED, PAID, COMPLETED]
      description: Current status of the order
    Order:
      type: object
      required:
        - product_barcode
        - quantity
        - price_per_unit
      properties:
        id:
          type: integer
          example: 1
        product_barcode:
          type: string
          example: "0123456789012"
        quantity:
          type: integer
          example: 100
        price_per_unit:
          type: number
          format: float
          example: 10.5
        status:
          $ref: '#/components/schemas/OrderStatus'
        issue_date:
          type: string
          format: date-time
          example: "2025-10-30T12:34:56Z"
...
paths:
...
  /orders:
    post:
      tags:
        - orders
      summary: Issue a new order
      description: Creates a new order for a given product. Requires **Administrator** or **ShopManager** role. 
        The new order is in the **ISSUED** state. If an id is provided in the request, it will be ignored.
      operationId: issueOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Order successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request (missing or invalid parameters)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Basrcode is required"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
        '404':
          description: Product Type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Product not found"
                name: "NotFoundError"
    get:
      tags:
        - orders
      summary: List all orders
      description: Returns all orders in the system. Requires **Administrator**, **ShopManager** role.
      operationId: listOrders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
  /orders/payfor:
    post:
      tags:
        - orders
      summary: Create and pay an order for a product
      description: |
        Creates a new order and immediately pays for it. Requires **Administrator** or **ShopManager** role. 
        The new order is in the **PAID** state. The balance is updated accordingly.
        If an id is provided in the request, it will be ignored.
      operationId: payOrderFor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Order successfully paid and created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request (missing parameters)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Barcode is required"
                name: "BadRequestError"
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Order not found"
                name: "NotFoundError"
        '421':
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 421
                message: "Insufficient balance for the operation"
                name: "BalanceError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
  /orders/{order_id}/pay:
    patch:
      tags:
        - orders
      summary: Pay for an existing issued order
      description: |
        Pays an existing ISSUED order. Requires **Administrator** or **ShopManager** role. 
        The order state is now **PAID**. The balance is updated accordingly.
      operationId: payOrder
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Order successfully paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '400':
          description: Invalid order ID or bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid order id"
                name: "BadRequestError"
        '404':
          description: Order not found or insufficient balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Order not found"
                name: "NotFoundError"
        '420':
          description: Invalid order state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 420
                message: "Order was not Issued"
                name: "InvalidStateError"
        '421':
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 421
                message: "Insufficient balance for the operation"
                name: "BalanceError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
  /orders/{order_id}/arrival:
    patch:
      tags:
        - orders
      summary: Record the arrival of an order
      description: Marks a **PAID** order as **COMPLETED** and updates product quantities. 
        Requires **Administrator** or **ShopManager** role.
      operationId: recordOrderArrival
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Order successfully marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '400':
          description: Invalid order ID or bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid order id"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
        '404':
          description: Order or product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Order not found"
                name: "NotFoundError"
        '420':
          description: Invalid order state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 420
                message: "order is not paid"
                name: "InvalidStateError"
        '500':
          description: General Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 500
                message: "Position must be set for an order before its arrival"
                name: "GeneralError"
...